IC-1446: Refactored postSessionAttendanceFeedbackPresenter to be decoupled from ActionPlanAppointment as we will use the same presentation for initial appointments
