+++ server/routes/service-provider/action-plan/appointment/post-session-feedback/attendance/actionPlanPostSessionAttendanceFeedbackPresenter.test.ts
@@ -1,0 +1,55 @@
+import actionPlanAppointmentFactory from '../../../../../../../testutils/factories/actionPlanAppointment'
+import deliusServiceUserFactory from '../../../../../../../testutils/factories/deliusServiceUser'
+import ActionPlanPostSessionAttendanceFeedbackPresenter from './actionPlanPostSessionAttendanceFeedbackPresenter'
+
+describe(ActionPlanPostSessionAttendanceFeedbackPresenter, () => {
+  describe('text', () => {
+    it('contains a title including the name of the service category and a subtitle, and the attendance questions', () => {
+      const appointment = actionPlanAppointmentFactory.build()
+      const serviceUser = deliusServiceUserFactory.build({ firstName: 'Alex' })
+      const presenter = new ActionPlanPostSessionAttendanceFeedbackPresenter(appointment, serviceUser)
+
+      expect(presenter.text).toMatchObject({
+        title: 'Add attendance feedback',
+        subTitle: 'Session details',
+        attendanceQuestion: 'Did Alex attend this session?',
+        attendanceQuestionHint: 'Select one option',
+        additionalAttendanceInformationLabel: "Add additional information about Alex's attendance:",
+      })
+    })
+    describe('when the session is a phone call', () => {
+      it('contains an attendance question to indicate the meeting was a phone call', () => {
+        const appointment = actionPlanAppointmentFactory.build({ appointmentDeliveryType: 'PHONE_CALL' })
+        const serviceUser = deliusServiceUserFactory.build({ firstName: 'Alex' })
+        const presenter = new ActionPlanPostSessionAttendanceFeedbackPresenter(appointment, serviceUser)
+        expect(presenter.text.attendanceQuestion).toEqual('Did Alex join this phone call?')
+      })
+    })
+    describe('when the session is a video call', () => {
+      it('contains an attendance question to indicate the meeting was a video call', () => {
+        const appointment = actionPlanAppointmentFactory.build({ appointmentDeliveryType: 'VIDEO_CALL' })
+        const serviceUser = deliusServiceUserFactory.build({ firstName: 'Alex' })
+        const presenter = new ActionPlanPostSessionAttendanceFeedbackPresenter(appointment, serviceUser)
+        expect(presenter.text.attendanceQuestion).toEqual('Did Alex join this video call?')
+      })
+    })
+    describe('when the session is an other location meeting', () => {
+      it('contains an attendance question to indicate the meeting was an in-person meeting', () => {
+        const appointment = actionPlanAppointmentFactory.build({ appointmentDeliveryType: 'IN_PERSON_MEETING_OTHER' })
+        const serviceUser = deliusServiceUserFactory.build({ firstName: 'Alex' })
+        const presenter = new ActionPlanPostSessionAttendanceFeedbackPresenter(appointment, serviceUser)
+        expect(presenter.text.attendanceQuestion).toEqual('Did Alex attend this in-person meeting?')
+      })
+    })
+    describe('when the session is an nps office location meeting', () => {
+      it('contains an attendance question to indicate the meeting was an in-person meeting', () => {
+        const appointment = actionPlanAppointmentFactory.build({
+          appointmentDeliveryType: 'IN_PERSON_MEETING_PROBATION_OFFICE',
+        })
+        const serviceUser = deliusServiceUserFactory.build({ firstName: 'Alex' })
+        const presenter = new ActionPlanPostSessionAttendanceFeedbackPresenter(appointment, serviceUser)
+        expect(presenter.text.attendanceQuestion).toEqual('Did Alex attend this in-person meeting?')
+      })
+    })
+  })
+})
